reestructured components folder
