+++ src/components/forms/CreateAlarmFormContainer.js
@@ -1,13 +1,12 @@
 import React, { useState } from "react";
-import { connect } from "react-redux";
 import AlarmForm from "./AlarmForm";
-import { editAlarm } from "../store/actions/alarms";
-import "../style/Forms.css";
+import { newAlarm } from "../../store/actions/alarms";
+import { connect } from "react-redux";
 
-function EditAlarmFormContainer(props) {
+function AlarmFormContainer(props) {
   const [alarm, setAlarm] = useState({
-    name: props.alarm.name,
-    time: props.alarm.time
+    name: "",
+    time: ""
   });
 
   const [dayOfTheWeek, setDayOfTheWeek] = useState({
@@ -23,28 +22,19 @@ function EditAlarmFormContainer(props) {
   const onSubmit = event => {
     event.preventDefault();
 
-    const time =
-      typeof alarm.time === "number"
-        ? alarm.time
-        : alarm.time
-            .split(":")
-            .map((string, index) => {
-              const timeInNumbers =
-                index === 0 ? parseInt(string) * 60 : parseInt(string);
-              return timeInNumbers;
-            })
-            .reduce((acc, minutes) => acc + minutes, 0);
+    const time = alarm.time
+      .split(":")
+      .map((string, index) => {
+        const timeInNumbers =
+          index === 0 ? parseInt(string) * 60 : parseInt(string);
+        return timeInNumbers;
+      })
+      .reduce((acc, minutes) => acc + minutes, 0);
 
     const days = Object.values(dayOfTheWeek).map((value, index) =>
       value ? index : null
     );
-    props
-      .editAlarm(props.plant.id, props.alarm.id, {
-        name: alarm.name,
-        time: time,
-        dayOfTheWeek: days
-      })
-      .then(props.onEdit);
+    props.newAlarm(alarm.name, time, days).then(props.onAdd);
   };
 
   const onChange = event => {
@@ -64,17 +54,22 @@ function EditAlarmFormContainer(props) {
   return (
     <div className="form ">
       <div className="card shadow-sm">
-        <p className="text-center">Edit information </p>
         <AlarmForm
           onSubmit={onSubmit}
           onChange={onChange}
           dayOnChange={dayOnChange}
           values={alarm}
-          button={"Update Alarm"}
+          button={"create new alarm"}
         />
       </div>
     </div>
   );
 }
 
-export default connect(null, { editAlarm })(EditAlarmFormContainer);
+function mapStateToProps(state) {
+  return { plant: state.plants.current };
+}
+
+const mapDispatchToProps = { newAlarm };
+
+export default connect(mapStateToProps, mapDispatchToProps)(AlarmFormContainer);
