+++ src/components/forms/CreateNoteFormContainer.js
@@ -1,23 +1,20 @@
 import React, { useState } from "react";
 import { connect } from "react-redux";
 import NoteForm from "./NoteForm";
-import { editNote } from "../store/actions/notes";
-import "../style/Forms.css";
+import { createNote } from "../../store/actions/notes";
+import "../../style/Forms.css";
 
-function EditNoteFormContainer(props) {
+function CreateNoteFormContainer(props) {
   const [note, setNote] = useState({
-    title: props.note.title,
-    text: props.note.text
+    title: "",
+    text: ""
   });
 
   const onSubmit = event => {
     event.preventDefault();
     props
-      .editNote(props.plant.id, props.note.id, {
-        title: note.title,
-        text: note.text
-      })
-      .then(props.onEdit);
+      .createNote(props.plant.id, { title: note.title, text: note.text })
+      .then(props.onAdd);
   };
 
   const onChange = event => {
@@ -26,20 +23,19 @@ function EditNoteFormContainer(props) {
       [event.target.name]: event.target.value
     });
   };
-
   return (
     <div className="form ">
       <div className="card shadow-sm">
-        <p className="text-center">Edit information </p>
+        <p className="text-center">Add information </p>
         <NoteForm
           onSubmit={onSubmit}
           onChange={onChange}
           values={note}
-          button={"Update Note"}
+          button={"Add Note"}
         />
       </div>
     </div>
   );
 }
 
-export default connect(null, { editNote })(EditNoteFormContainer);
+export default connect(null, { createNote })(CreateNoteFormContainer);
