+++ src/components/forms/EditAlarmFormContainer.js
@@ -1,12 +1,13 @@
 import React, { useState } from "react";
-import AlarmForm from "./AlarmForm";
-import { newAlarm } from "../store/actions/alarms";
 import { connect } from "react-redux";
+import AlarmForm from "./AlarmForm";
+import { editAlarm } from "../../store/actions/alarms";
+import "../../style/Forms.css";
 
-function AlarmFormContainer(props) {
+function EditAlarmFormContainer(props) {
   const [alarm, setAlarm] = useState({
-    name: "",
-    time: ""
+    name: props.alarm.name,
+    time: props.alarm.time
   });
 
   const [dayOfTheWeek, setDayOfTheWeek] = useState({
@@ -22,19 +23,28 @@ function AlarmFormContainer(props) {
   const onSubmit = event => {
     event.preventDefault();
 
-    const time = alarm.time
-      .split(":")
-      .map((string, index) => {
-        const timeInNumbers =
-          index === 0 ? parseInt(string) * 60 : parseInt(string);
-        return timeInNumbers;
-      })
-      .reduce((acc, minutes) => acc + minutes, 0);
+    const time =
+      typeof alarm.time === "number"
+        ? alarm.time
+        : alarm.time
+            .split(":")
+            .map((string, index) => {
+              const timeInNumbers =
+                index === 0 ? parseInt(string) * 60 : parseInt(string);
+              return timeInNumbers;
+            })
+            .reduce((acc, minutes) => acc + minutes, 0);
 
     const days = Object.values(dayOfTheWeek).map((value, index) =>
       value ? index : null
     );
-    props.newAlarm(alarm.name, time, days).then(props.onAdd);
+    props
+      .editAlarm(props.plant.id, props.alarm.id, {
+        name: alarm.name,
+        time: time,
+        dayOfTheWeek: days
+      })
+      .then(props.onEdit);
   };
 
   const onChange = event => {
@@ -54,22 +64,17 @@ function AlarmFormContainer(props) {
   return (
     <div className="form ">
       <div className="card shadow-sm">
+        <p className="text-center">Edit information </p>
         <AlarmForm
           onSubmit={onSubmit}
           onChange={onChange}
           dayOnChange={dayOnChange}
           values={alarm}
-          button={"create new alarm"}
+          button={"Update Alarm"}
         />
       </div>
     </div>
   );
 }
 
-function mapStateToProps(state) {
-  return { plant: state.plants.current };
-}
-
-const mapDispatchToProps = { newAlarm };
-
-export default connect(mapStateToProps, mapDispatchToProps)(AlarmFormContainer);
+export default connect(null, { editAlarm })(EditAlarmFormContainer);
