+++ src/components/pages/AllNotesPage.js
@@ -1,9 +1,9 @@
 import React, { useEffect } from "react";
-import NoteCard from "./NoteCard";
-import { getUserNotes } from "../store/actions/notes";
+import NoteCard from "../presentationals/NoteCard";
+import { getUserNotes } from "../../store/actions/notes";
 import { connect } from "react-redux";
 import { Redirect } from "react-router-dom";
-import "../style/AllAlarmsPage.css";
+import "../../style/AllAlarmsPage.css";
 
 function AllNotesPage(props) {
   const { getUserNotes } = props;
