+++ src/components/presentationals/Navbar.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import { connect } from "react-redux";
-import { logout } from "../store/actions/user";
-import "../style/Navbar.css";
+import { logout } from "../../store/actions/user";
+import "../../style/Navbar.css";
 
 function Navbar(props) {
   const onLogout = () => {
